refactor(compact): decompose compact.rs (1363 → 385 lines) into submodules
Pure, behavior-preserving split of the oversized compact.rs into cohesive
compact/ submodules. Every public item is re-exported so the canonical
`crate::compact::X` paths used across the workspace are unchanged — no call
site outside the module moved.

  compact/record.rs     CompactRecord + NTFS metafile-name allowlist   (189)
  compact/children.rs   ChildrenIndex (CSR parent→children)            (111)
  compact/extension.rs  ExtensionIndex (CSR ext_id→records)            (102)
  compact/path_len.rs   compute_path_lengths + Phase-1 incremental fns (214)
  compact/builder.rs    build_compact_index + ADS/links/shrink/upcase  (422)
  compact.rs            DriveCompactIndex + HeapReport + impl + re-exports (385)

compact.rs drops off the file-size exception list (was "13 over"; now 385,
well under 800). 867/867 uffs-core tests pass unchanged (identical count
pre/post — proves a pure move); clippy -D warnings, rustdoc -D warnings,
lint-ci-windows all clean.

This also tidies the tree for Phase 2b: the compact() method + apply
delta-population drop cleanly into builder.rs / a slim compact.rs rather
than a 1363-line file.
