+++ crates/uffs-core/src/compact.rs
@@ -9,13 +9,11 @@
 //! `.uffs` cache file.
 //!
 //! See `docs/architecture/COMPACT_INDEX_DESIGN.md` for the full design.
-//! Exception: `file_size_policy` — core data structures + builder, tightly
-//! coupled.
-
-use std::time::Instant;
-
-use rayon::prelude::*;
-use uffs_mft::index::MftIndex;
+//!
+//! This module owns [`DriveCompactIndex`] (the loaded drive + its search choke
+//! points) and re-exports the row type, the CSR indexes, path-length
+//! computation, and the MFT→compact builder from focused submodules
+//! (`record`, `children`, `extension`, `path_len`, `builder`, `delta`).
 
 use crate::bloom::Bloom;
 pub use crate::compact_loader::apply_usn_patch;
@@ -31,388 +29,24 @@ use crate::trigram::TrigramIndex;
 /// Mutable delta overlay over the immutable base CSR indexes (Phase 2+).
 pub mod delta;
 
+// File-size decomposition: the row type, the CSR indexes, path-length
+// computation, and the MFT→compact builder live in focused submodules.  Every
+// public item is re-exported below so the canonical `crate::compact::X` paths
+// (used across the workspace) are unchanged.
+mod builder;
+mod children;
+mod extension;
+mod path_len;
+mod record;
+
+pub use builder::build_compact_index;
+pub(crate) use builder::{INDEX_TTL_SECONDS, resolve_case_fold};
+pub use children::ChildrenIndex;
 pub use delta::IndexDelta;
-
-/// Compact per-record data for in-memory search, filter, and sort.
-///
-/// 80 bytes per record (76 data + 4 explicit tail padding).
-/// Derives `bytemuck::Pod` + `Zeroable` so the entire record array can be
-/// serialized/deserialized as a single bulk `memcpy` — no per-field encoding.
-#[derive(Debug, Clone, Copy, Default, bytemuck::Pod, bytemuck::Zeroable)]
-#[repr(C)]
-pub struct CompactRecord {
-    // ── u64 fields first (8-byte aligned) ─────────────────────────
-    /// Logical file size in bytes.
-    pub size: u64,
-    /// Allocated size on disk in bytes ("Size on Disk" column).
-    pub allocated: u64,
-    /// Sum of logical file sizes in entire subtree.
-    pub treesize: u64,
-    /// Sum of allocated sizes in entire subtree.
-    pub tree_allocated: u64,
-    /// Creation time (Unix microseconds).
-    pub created: i64,
-    /// Last write time (Unix microseconds).
-    pub modified: i64,
-    /// Last access time (Unix microseconds).
-    pub accessed: i64,
-
-    // ── u32 fields (4-byte aligned) ───────────────────────────────
-    /// Byte offset into the names blob.
-    pub name_offset: u32,
-    /// Raw NTFS `FILE_ATTRIBUTE_*` flags.
-    pub flags: u32,
-    /// Index into the compact array of the parent directory.
-    /// `u32::MAX` = root or orphan.
-    pub parent_idx: u32,
-    /// Count of all descendants in subtree. 0 for files.
-    pub descendants: u32,
-
-    // ── u16 fields (2-byte aligned) ───────────────────────────────
-    /// UTF-8 byte length of the filename.
-    pub name_len: u16,
-    /// Interned extension ID (0 = no extension).
-    pub extension_id: u16,
-    /// Full path length in UTF-8 bytes (e.g. `C:\Windows\System32\cmd.exe` =
-    /// 28). Precomputed at index build time via top-down parent-chain walk.
-    /// Saturates at `u16::MAX` (65 535) for extremely deep paths.
-    pub path_len: u16,
-
-    /// First byte of the filename (e.g. `b'$'` for NTFS metafiles).
-    ///
-    /// Cached here as a cheap hot-path *gate*: only `$`-prefixed records can be
-    /// NTFS metafiles, so [`is_system_metafile`](Self::is_system_metafile) can
-    /// reject virtually every record with one sequential field read instead of
-    /// a random cache-miss into the names arena.  The handful of `$`-prefixed
-    /// candidates then pay one arena lookup for the authoritative name check.
-    pub name_first_byte: u8,
-
-    /// Explicit tail padding for 8-byte struct alignment.
-    /// Required by `bytemuck::Pod` — no implicit padding allowed.
-    #[expect(
-        clippy::pub_underscore_fields,
-        reason = "bytemuck Pod requires all fields same visibility"
-    )]
-    pub _pad: [u8; 1],
-}
-
-/// The fixed set of reserved NTFS metafile names: the `$`-prefixed records at
-/// reserved FRS 0–15 and under the `$Extend` directory.  An NTFS volume can
-/// only ever contain *these* specific metafiles.
-///
-/// Any *other* `$`-prefixed name — `$Recycle.Bin`, `$PatchCache`,
-/// `$WinREAgent`, the `WinSxS` `$$_*.cdf-ms` filemaps, or a user file literally
-/// named `$foo` — is an ordinary file that file managers and tools like
-/// Everything display. Classifying those as metafiles is exactly the bug
-/// `--hide-system` had.
-///
-/// Matched case-insensitively: NTFS itself is case-insensitive, and these
-/// canonical names are occasionally surfaced with varied casing.
-pub(crate) const NTFS_METAFILE_NAMES: &[&str] = &[
-    // Reserved FRS 0–11 (volume root metafiles)
-    "$MFT",
-    "$MFTMirr",
-    "$LogFile",
-    "$Volume",
-    "$AttrDef",
-    "$Bitmap",
-    "$Boot",
-    "$BadClus",
-    "$Secure",
-    "$UpCase",
-    "$Extend",
-    // `$Extend` directory children
-    "$ObjId",
-    "$Quota",
-    "$Reparse",
-    "$UsnJrnl",
-    "$RmMetadata",
-    "$Deleted",
-    // `$Extend\$RmMetadata` children
-    "$Repair",
-    "$Tops",
-    "$TxfLog",
-    "$Txf",
-];
-
-/// Returns whether `name` is one of the reserved `NTFS_METAFILE_NAMES`
-/// (a crate-private allowlist, so no intra-doc link from this public item).
-///
-/// Real metafiles are already excluded from the compact index at build time
-/// (`build_compact_index` drops them via `PathResolver` FRS-validity, not by
-/// name).  This exact-name check is the *authoritative* classifier for the
-/// `--hide-system` filter, so it can never misclassify an ordinary
-/// `$`-prefixed file as a metafile.
-#[must_use]
-#[inline]
-pub fn is_ntfs_metafile_name(name: &str) -> bool {
-    NTFS_METAFILE_NAMES
-        .iter()
-        .any(|reserved| name.eq_ignore_ascii_case(reserved))
-}
-
-impl CompactRecord {
-    /// Directory flag bit in raw NTFS `FILE_ATTRIBUTE_DIRECTORY`.
-    const DIRECTORY_BIT: u32 = 0x0010;
-
-    /// Returns `true` if this record is a directory.
-    #[inline]
-    #[must_use]
-    pub const fn is_directory(self) -> bool {
-        self.flags & Self::DIRECTORY_BIT != 0
-    }
-
-    /// Returns `true` if this record is one of the reserved NTFS metafiles
-    /// (`$MFT`, `$LogFile`, `$Bitmap`, `$Secure`, the `$Extend` family, …).
-    ///
-    /// The cached [`name_first_byte`](Self::name_first_byte) field is a cheap
-    /// gate: every metafile name starts with `$`, and `$`-prefixed records are
-    /// a vanishing fraction of an index, so this rejects virtually every record
-    /// with a single byte comparison and only touches the names arena for the
-    /// handful of `$`-prefixed candidates.  The arena lookup is *required* for
-    /// correctness, because an ordinary file may also start with `$`
-    /// (`$Recycle.Bin`, `$PatchCache`, the `WinSxS` `$$_*.cdf-ms` filemaps) —
-    /// those are NOT metafiles and must not be hidden by `--hide-system`.
-    /// See [`is_ntfs_metafile_name`].
-    #[inline]
-    #[must_use]
-    pub fn is_system_metafile(&self, names: &[u8]) -> bool {
-        self.name_first_byte == b'$' && is_ntfs_metafile_name(self.name(names))
-    }
-
-    /// Get the name from a names blob as a **lossy `&str` view**.
-    ///
-    /// Valid-UTF-8 names (the common case) are returned verbatim; an ill-formed
-    /// (surrogate-bearing) name stored as WTF-8 returns `""` for display. Use
-    /// [`Self::name_bytes`] for the lossless bytes that exact/substring search
-    /// matches against, so a file with an ill-formed name stays findable
-    /// (WI-4.4).
-    #[inline]
-    #[must_use]
-    pub fn name<'a>(&self, names: &'a [u8]) -> &'a str {
-        core::str::from_utf8(self.name_bytes(names)).unwrap_or("")
-    }
-
-    /// Get the name's **raw bytes** (WTF-8) from a names blob — the lossless
-    /// accessor.
-    ///
-    /// Returns exactly the stored bytes, including the byte-faithful encoding
-    /// of an ill-formed NTFS name (unpaired surrogates). This is what makes
-    /// every file matchable/findable by its true name regardless of UTF-8
-    /// well-formedness (WI-4.4). Returns `&[]` for an out-of-range slice.
-    #[inline]
-    #[must_use]
-    pub fn name_bytes<'a>(&self, names: &'a [u8]) -> &'a [u8] {
-        let start = self.name_offset as usize;
-        let end = start.saturating_add(self.name_len as usize);
-        names.get(start..end).unwrap_or(&[])
-    }
-}
-
-// Compile-time size assertion.
-const _: () = assert!(
-    size_of::<CompactRecord>() == 80,
-    "CompactRecord must be exactly 80 bytes"
-);
-
-/// Children index in CSR (Compressed Sparse Row) layout.
-///
-/// `children(i)` returns the compact indices of record i's children as
-/// a contiguous `&[u32]` slice.  The CSR layout avoids per-record `Vec`
-/// allocations and enables bulk serialization/deserialization.
-#[derive(Clone)]
-pub struct ChildrenIndex {
-    /// CSR offsets — one per record + sentinel.  Length = `record_count` + 1.
-    /// Children of record `i` are `values[offsets[i]..offsets[i+1]]`.
-    offsets: Vec<u32>,
-    /// Flat array of all child indices.
-    values: Vec<u32>,
-}
-
-impl ChildrenIndex {
-    /// Total heap capacity (offsets + values) in bytes.
-    #[must_use]
-    pub const fn heap_size_bytes(&self) -> usize {
-        self.offsets.capacity() * size_of::<u32>() + self.values.capacity() * size_of::<u32>()
-    }
-
-    /// Build from `CompactRecord::parent_idx` in two passes (count + scatter).
-    #[must_use]
-    pub fn build(records: &[CompactRecord]) -> Self {
-        // Count children per parent
-        let mut counts = vec![0_u32; records.len()];
-        for rec in records {
-            let parent = rec.parent_idx;
-            if parent != u32::MAX
-                && let Some(cnt) = counts.get_mut(parent as usize)
-            {
-                *cnt += 1;
-            }
-        }
-
-        // Prefix-sum → offsets
-        let mut offsets = Vec::with_capacity(records.len() + 1);
-        let mut running = 0_u32;
-        for &cnt in &counts {
-            offsets.push(running);
-            running = running.saturating_add(cnt);
-        }
-        offsets.push(running);
-
-        // Scatter children into values
-        let mut values = vec![0_u32; running as usize];
-        let mut write_pos = offsets.clone();
-        for (idx, rec) in records.iter().enumerate() {
-            let parent = rec.parent_idx;
-            if parent != u32::MAX
-                && let Some(pos) = write_pos.get_mut(parent as usize)
-                && let Some(slot) = values.get_mut(*pos as usize)
-            {
-                let child_idx = uffs_mft::len_to_u32(idx);
-                *slot = child_idx;
-                *pos += 1;
-            }
-        }
-
-        Self { offsets, values }
-    }
-
-    /// Construct directly from pre-built CSR arrays (cache deserialization).
-    #[must_use]
-    pub const fn from_csr(offsets: Vec<u32>, values: Vec<u32>) -> Self {
-        Self { offsets, values }
-    }
-
-    /// Borrow the CSR components for serialization.
-    #[must_use]
-    pub(crate) fn as_csr(&self) -> (&[u32], &[u32]) {
-        (&self.offsets, &self.values)
-    }
-
-    /// Return the children of record `idx` as a contiguous slice.
-    #[must_use]
-    pub fn get(&self, idx: usize) -> &[u32] {
-        let start = self.offsets.get(idx).copied().unwrap_or(0) as usize;
-        let end = self.offsets.get(idx + 1).copied().unwrap_or(0) as usize;
-        self.values.get(start..end).unwrap_or(&[])
-    }
-
-    /// Total number of child entries across all records.
-    #[must_use]
-    pub const fn total_children(&self) -> usize {
-        self.values.len()
-    }
-
-    /// Number of records tracked (one slot per record).
-    #[must_use]
-    pub const fn record_count(&self) -> usize {
-        self.offsets.len().saturating_sub(1)
-    }
-
-    /// Create an empty children index.
-    #[must_use]
-    pub fn empty() -> Self {
-        Self {
-            offsets: vec![0],
-            values: Vec::new(),
-        }
-    }
-}
-
-/// Extension inverted index: `extension_id → &[u32]` (record indices).
-///
-/// CSR layout identical to `ChildrenIndex`.  Built once at load time in a
-/// single O(N) pass so `--ext rs` queries can iterate only matching records
-/// instead of scanning all 25M entries.
-#[derive(Clone)]
-pub struct ExtensionIndex {
-    /// CSR offsets — length = `max_ext_id` + 2 (one per `ext_id` + sentinel).
-    offsets: Vec<u32>,
-    /// Flat array of record indices, grouped by `extension_id`.
-    values: Vec<u32>,
-}
-
-impl ExtensionIndex {
-    /// Total heap capacity (offsets + values) in bytes.
-    #[must_use]
-    pub const fn heap_size_bytes(&self) -> usize {
-        self.offsets.capacity() * size_of::<u32>() + self.values.capacity() * size_of::<u32>()
-    }
-
-    /// Build from compact records in two passes (count + scatter).
-    #[must_use]
-    pub fn build(records: &[CompactRecord]) -> Self {
-        // Find the maximum extension_id to size the offsets array.
-        let max_id = records
-            .iter()
-            .map(|rec| rec.extension_id)
-            .max()
-            .unwrap_or(0) as usize;
-
-        // Pass 1: count records per extension_id.
-        let mut counts = vec![0_u32; max_id + 1];
-        for rec in records {
-            if rec.name_len == 0 {
-                continue;
-            }
-            if let Some(cnt) = counts.get_mut(rec.extension_id as usize) {
-                *cnt += 1;
-            }
-        }
-
-        // Prefix-sum → offsets.
-        let mut offsets = Vec::with_capacity(max_id + 2);
-        let mut running = 0_u32;
-        for &cnt in &counts {
-            offsets.push(running);
-            running = running.saturating_add(cnt);
-        }
-        offsets.push(running);
-
-        // Pass 2: scatter record indices into values.
-        let mut values = vec![0_u32; running as usize];
-        let mut write_pos = offsets.clone();
-        for (idx, rec) in records.iter().enumerate() {
-            if rec.name_len == 0 {
-                continue;
-            }
-            let eid = rec.extension_id as usize;
-            if let Some(pos) = write_pos.get_mut(eid)
-                && let Some(slot) = values.get_mut(*pos as usize)
-            {
-                let idx_u32 = uffs_mft::len_to_u32(idx);
-                *slot = idx_u32;
-                *pos += 1;
-            }
-        }
-
-        Self { offsets, values }
-    }
-
-    /// Return record indices for the given `extension_id`.
-    #[must_use]
-    pub fn get(&self, ext_id: u16) -> &[u32] {
-        let eid = ext_id as usize;
-        let start = self.offsets.get(eid).copied().unwrap_or(0) as usize;
-        let end = self.offsets.get(eid + 1).copied().unwrap_or(0) as usize;
-        self.values.get(start..end).unwrap_or(&[])
-    }
-
-    /// Create an empty extension index.
-    #[must_use]
-    pub fn empty() -> Self {
-        Self {
-            offsets: vec![0],
-            values: Vec::new(),
-        }
-    }
-
-    /// Total number of indexed record entries.
-    #[must_use]
-    pub const fn total_entries(&self) -> usize {
-        self.values.len()
-    }
-}
+pub use extension::ExtensionIndex;
+pub(crate) use path_len::{PathChange, compute_path_lengths, update_path_lengths_incremental};
+pub(crate) use record::NTFS_METAFILE_NAMES;
+pub use record::{CompactRecord, is_ntfs_metafile_name};
 
 /// A loaded drive with compact index.
 #[derive(Clone)]
@@ -735,615 +369,6 @@ impl DriveCompactIndex {
         }
         self.ext_names.push(normalized.into_boxed_str());
         new_id
-    }
-}
-
-/// Expand alternate data streams (ADS) for a single record, producing the
-/// name × stream cross product as extra `CompactRecord` entries.
-#[expect(
-    clippy::single_call_fn,
-    reason = "Extracted to keep expand_links_and_ads under the too_many_lines limit"
-)]
-fn expand_ads_streams(
-    index: &MftIndex,
-    record: &uffs_mft::index::FileRecord,
-    resolve_parent: &dyn Fn(uffs_mft::ParentFrs, uffs_mft::Frs) -> u32,
-    names: &mut Vec<u8>,
-    extra: &mut Vec<CompactRecord>,
-) {
-    // Collect all names for this record (primary + hardlinks).
-    let mut all_names: Vec<(&str, u32)> = Vec::new();
-    let primary_name = index.get_name(record.first_name.name);
-    if !primary_name.is_empty() {
-        let pid = resolve_parent(record.first_name.parent_frs, record.frs);
-        all_names.push((primary_name, pid));
-    }
-    if record.name_count > 1 {
-        let mut le = record.first_name.next_entry;
-        while le != uffs_mft::NO_ENTRY {
-            let Some(lnk) = index.links.get(le as usize) else {
-                break;
-            };
-            let ln = index.get_name(lnk.name);
-            if !ln.is_empty() {
-                let lp = resolve_parent(lnk.parent_frs, record.frs);
-                all_names.push((ln, lp));
-            }
-            le = lnk.next_entry;
-        }
-    }
-
-    // Walk output streams (skip default $DATA at head of chain).
-    let mut se = record.first_stream.next_entry;
-    while se != uffs_mft::NO_ENTRY {
-        let Some(stream) = index.streams.get(se as usize) else {
-            break;
-        };
-        if stream.is_output_stream() {
-            let sn = index.stream_name(stream);
-            if !sn.is_empty() {
-                for &(base_name, parent_idx) in &all_names {
-                    let combined = format!("{base_name}:{sn}");
-                    let name_offset = uffs_mft::len_to_u32(names.len());
-                    let name_len = uffs_mft::len_to_u16(combined.len());
-                    names.extend_from_slice(combined.as_bytes());
-
-                    extra.push(CompactRecord {
-                        size: stream.size.length,
-                        allocated: stream.size.allocated,
-                        treesize: 0,
-                        tree_allocated: 0,
-                        created: record.stdinfo.created,
-                        modified: record.stdinfo.modified,
-                        accessed: record.stdinfo.accessed,
-                        name_offset,
-                        flags: record.stdinfo.flags,
-                        parent_idx,
-                        descendants: 0,
-                        name_len,
-                        extension_id: 0,
-                        path_len: 0,
-                        name_first_byte: combined.as_bytes().first().copied().unwrap_or(0),
-                        _pad: [0; 1],
-                    });
-                }
-            }
-        }
-        se = stream.next_entry;
-    }
-}
-
-/// Resolve a typed `ParentFrs` (vs an own typed `Frs`) into a compact-record
-/// index, returning `u32::MAX` for the "no real parent" cases (self-reference,
-/// `NO_ENTRY` sentinel, or root).
-///
-/// Extracted as a free helper so the typed `ParentFrs`/`Frs` signature is
-/// enforced at every call site AND so `build_compact_index` stays under
-/// the clippy `too_many_lines` budget.
-#[expect(
-    clippy::single_call_fn,
-    reason = "Wrapped by a closure in build_compact_index; kept free-standing \
-              for clippy::too_many_lines budget headroom"
-)]
-fn resolve_parent_compact_idx(
-    index: &MftIndex,
-    parent_frs: uffs_mft::ParentFrs,
-    own_frs: uffs_mft::Frs,
-) -> u32 {
-    let parent = parent_frs.as_frs();
-    if parent == own_frs || parent_frs.raw() == u64::from(uffs_mft::NO_ENTRY) || parent.is_root() {
-        return u32::MAX;
-    }
-    let parent_usize = uffs_mft::frs_to_usize(parent.raw());
-    index
-        .frs_to_idx
-        .get(parent_usize)
-        .copied()
-        .filter(|&idx| idx != uffs_mft::NO_ENTRY)
-        .unwrap_or(u32::MAX)
-}
-
-/// Expand hardlinks and ADS into additional `CompactRecord` entries.
-///
-/// Phase 2 (hardlinks): for each valid record with `name_count > 1`, walks the
-/// link chain and creates additional records with alternate name/parent.
-///
-/// Phase 3 (ADS): delegates to [`expand_ads_streams`] for each valid record
-/// with `stream_count > 1`.
-#[expect(
-    clippy::single_call_fn,
-    reason = "Extracted to keep build_compact_index under the too_many_lines limit"
-)]
-fn expand_links_and_ads(
-    index: &MftIndex,
-    resolver: &uffs_mft::index::PathResolver,
-    resolve_parent: &dyn Fn(uffs_mft::ParentFrs, uffs_mft::Frs) -> u32,
-    names: &mut Vec<u8>,
-) -> Vec<CompactRecord> {
-    let mut extra: Vec<CompactRecord> = Vec::new();
-
-    for (idx, record) in index.records.iter().enumerate() {
-        if !resolver.is_valid_idx(idx) {
-            continue;
-        }
-
-        // Phase 2: hardlink expansion.
-        if record.name_count > 1 {
-            let mut link_entry = record.first_name.next_entry;
-            while link_entry != uffs_mft::NO_ENTRY {
-                let Some(link) = index.links.get(link_entry as usize) else {
-                    break;
-                };
-                let link_parent = resolve_parent(link.parent_frs, record.frs);
-                extra.push(CompactRecord {
-                    size: record.first_stream.size.length,
-                    allocated: record.first_stream.size.allocated,
-                    treesize: record.treesize,
-                    tree_allocated: record.tree_allocated,
-                    created: record.stdinfo.created,
-                    modified: record.stdinfo.modified,
-                    accessed: record.stdinfo.accessed,
-                    name_offset: link.name.offset,
-                    flags: record.stdinfo.flags,
-                    parent_idx: link_parent,
-                    descendants: record.descendants,
-                    name_len: link.name.length(),
-                    extension_id: link.name.extension_id(),
-                    path_len: 0,
-                    name_first_byte: names.get(link.name.offset as usize).copied().unwrap_or(0),
-                    _pad: [0; 1],
-                });
-                link_entry = link.next_entry;
-            }
-        }
-
-        // Phase 3: ADS expansion (name × stream cross product).
-        if record.stream_count > 1 {
-            expand_ads_streams(index, record, resolve_parent, names, &mut extra);
-        }
-    }
-    extra
-}
-
-/// Compute `path_len` (in **characters**, not bytes) for every record
-/// via top-down BFS.
-///
-/// Root entries (`parent_idx == u32::MAX`) get
-/// `path_len = 2 + 1 + name_chars` (e.g. `"C:\" + name`), and children
-/// accumulate `parent.path_len + 1 (separator) + name_chars`.
-/// Saturates at `u16::MAX` (65 535) for extremely deep paths.
-///
-/// Character counting matches `str::chars().count()` so the precomputed
-/// value agrees with the display-row path-length filter.
-pub(crate) fn compute_path_lengths(
-    records: &mut [CompactRecord],
-    names: &[u8],
-    drive_letter: uffs_mft::platform::DriveLetter,
-) {
-    // Drive prefix in characters: the letter (1 char) + colon (1 char) = 2.
-    // `DriveLetter` is ASCII A–Z by construction (validated in
-    // `DriveLetter::parse`), so the previous runtime `debug_assert!`
-    // is now a tautology and was removed.  The arithmetic only cares
-    // about "1 letter char + 1 colon".
-    let _: uffs_mft::platform::DriveLetter = drive_letter;
-    let drive_prefix_chars: u32 = 1 /* letter */ + 1 /* ':' */;
-
-    // Build forward adjacency list (parent → children) for top-down BFS.
-    let record_count = records.len();
-    let mut children_of: Vec<Vec<u32>> = vec![Vec::new(); record_count];
-    let mut roots: Vec<u32> = Vec::new();
-
-    for (idx, rec) in records.iter().enumerate() {
-        let pi = rec.parent_idx;
-        if pi == u32::MAX {
-            roots.push(uffs_mft::len_to_u32(idx));
-        } else if let Some(siblings) = children_of.get_mut(pi as usize) {
-            siblings.push(uffs_mft::len_to_u32(idx));
-        }
-    }
-
-    // BFS from roots.
-    let mut queue = alloc::collections::VecDeque::with_capacity(roots.len());
-    for &root in &roots {
-        let Some(rec) = records.get(root as usize) else {
-            continue;
-        };
-        let name_chars = name_char_count(rec, names);
-        let pl = if name_chars == 0 {
-            // Drive root directory: "C:\"
-            drive_prefix_chars + 1
-        } else {
-            // Top-level file/dir: "C:\<name>"
-            drive_prefix_chars + 1 + name_chars
-        };
-        if let Some(slot) = records.get_mut(root as usize) {
-            slot.path_len = uffs_mft::len_to_u16(pl as usize);
-        }
-        queue.push_back(root);
-    }
-
-    while let Some(idx) = queue.pop_front() {
-        let parent_pl = records
-            .get(idx as usize)
-            .map_or(0, |rec| u32::from(rec.path_len));
-        let children: Vec<u32> = children_of
-            .get(idx as usize)
-            .map_or_else(Vec::new, Clone::clone);
-        for &child in &children {
-            let child_chars = records
-                .get(child as usize)
-                .map_or(0, |rec| name_char_count(rec, names));
-            // path = parent_path + "\" + name
-            let pl = parent_pl.saturating_add(1).saturating_add(child_chars);
-            if let Some(slot) = records.get_mut(child as usize) {
-                slot.path_len = uffs_mft::len_to_u16(pl as usize);
-            }
-            queue.push_back(child);
-        }
-    }
-}
-
-/// A record whose `path_len` a USN apply must refresh, plus whether the change
-/// can shift its whole subtree.
-///
-/// Phase 1 of incremental-index-maintenance (design doc §5.5): instead of the
-/// O(total) [`compute_path_lengths`] BFS every apply, refresh only the records
-/// a batch touched.  A **directory rename** moves every descendant's path by a
-/// constant Δ, so `subtree` requests the descendant walk; creates and file
-/// renames are a single O(1) refresh.
-#[derive(Debug, Clone, Copy)]
-pub(crate) struct PathChange {
-    /// Compact index of the created / renamed record.
-    pub idx: u32,
-    /// `true` for a directory rename (propagate Δ to descendants); `false` for
-    /// creates and file renames (refresh this record only).
-    pub subtree: bool,
-}
-
-/// Refresh `path_len` for only the records a USN batch touched, instead of the
-/// O(total-records) [`compute_path_lengths`] BFS — the Phase-1 lever of
-/// incremental-index-maintenance (design doc §5.5).
-///
-/// `children` must be the **freshly rebuilt** CSR so a directory rename can
-/// walk its subtree.  Caller falls back to the full [`compute_path_lengths`]
-/// for cold loads and for batches large enough that incremental loses
-/// (see the threshold in `compact_loader/rebuild.rs`).
-pub(crate) fn update_path_lengths_incremental(
-    records: &mut [CompactRecord],
-    names: &[u8],
-    drive_letter: uffs_mft::platform::DriveLetter,
-    children: &ChildrenIndex,
-    changed: &[PathChange],
-) {
-    // `DriveLetter` is ASCII A–Z by construction, so the drive prefix is always
-    // "X:" = 2 chars (matches `compute_path_lengths`).
-    let _: uffs_mft::platform::DriveLetter = drive_letter;
-    let drive_prefix_chars: u32 = 1 /* letter */ + 1 /* ':' */;
-
-    for change in changed {
-        let idx = change.idx as usize;
-        let Some(rec) = records.get(idx) else {
-            continue;
-        };
-        // Skip a slot tombstoned within the same batch (create then delete):
-        // `apply_delete` set name_len=0 + parent=MAX.
-        if rec.name_len == 0 && rec.parent_idx == u32::MAX {
-            continue;
-        }
-        let old_pl = u32::from(rec.path_len);
-        let new_pl = path_len_from_parent(records, names, drive_prefix_chars, change.idx);
-        if let Some(slot) = records.get_mut(idx) {
-            slot.path_len = uffs_mft::len_to_u16(new_pl as usize);
-        }
-        if change.subtree {
-            let delta = i64::from(new_pl) - i64::from(old_pl);
-            if delta != 0 {
-                shift_subtree_path_len(records, children, change.idx, delta);
-            }
-        }
-    }
-}
-
-/// `path_len` for `idx` from its (current) parent's `path_len` + own name —
-/// the per-node arithmetic of [`compute_path_lengths`]'s BFS, in isolation.
-fn path_len_from_parent(
-    records: &[CompactRecord],
-    names: &[u8],
-    drive_prefix_chars: u32,
-    idx: u32,
-) -> u32 {
-    let Some(rec) = records.get(idx as usize) else {
-        return 0;
-    };
-    let name_chars = name_char_count(rec, names);
-    if rec.parent_idx == u32::MAX {
-        // Root level: "C:\" (no name) or "C:\<name>".
-        if name_chars == 0 {
-            drive_prefix_chars.saturating_add(1)
-        } else {
-            drive_prefix_chars
-                .saturating_add(1)
-                .saturating_add(name_chars)
-        }
-    } else {
-        let parent_pl = records
-            .get(rec.parent_idx as usize)
-            .map_or(0, |parent| u32::from(parent.path_len));
-        parent_pl.saturating_add(1).saturating_add(name_chars)
-    }
-}
-
-/// Add `delta` to every descendant of `root`'s `path_len` (a directory rename
-/// shifts each descendant's full path by the same amount).  Iterative DFS over
-/// the children CSR — pure arithmetic, no name/string walk.
-fn shift_subtree_path_len(
-    records: &mut [CompactRecord],
-    children: &ChildrenIndex,
-    root: u32,
-    delta: i64,
-) {
-    let mut stack: Vec<u32> = children.get(root as usize).to_vec();
-    while let Some(idx) = stack.pop() {
-        if let Some(rec) = records.get_mut(idx as usize) {
-            let shifted = i64::from(rec.path_len)
-                .saturating_add(delta)
-                .clamp(0, i64::from(u16::MAX));
-            rec.path_len = u16::try_from(shifted).unwrap_or(u16::MAX);
-        }
-        stack.extend_from_slice(children.get(idx as usize));
-    }
-}
-
-/// Count the number of Unicode characters in a record's filename.
-///
-/// Falls back to `name_len` (byte count) if the name slice is not valid
-/// UTF-8 — this is correct for ASCII names and a safe upper bound
-/// otherwise.
-fn name_char_count(rec: &CompactRecord, names: &[u8]) -> u32 {
-    let start = rec.name_offset as usize;
-    let end = start + rec.name_len as usize;
-    names
-        .get(start..end)
-        .and_then(|slice| core::str::from_utf8(slice).ok())
-        .map_or_else(
-            || u32::from(rec.name_len),
-            |name| uffs_mft::len_to_u32(name.chars().count()),
-        )
-}
-
-/// Build a `DriveCompactIndex` from a loaded `MftIndex`.
-///
-/// Returns `(DriveCompactIndex, compact_build_ms, trigram_build_ms)`.
-#[must_use]
-pub fn build_compact_index(
-    drive_letter: uffs_mft::platform::DriveLetter,
-    index: &MftIndex,
-) -> (DriveCompactIndex, u128, u128) {
-    use uffs_mft::index::PathResolver;
-
-    let compact_start = Instant::now();
-
-    // Build path resolver to determine which records are valid.
-    // This filters out system metafiles (FRS 0-15 except root) and
-    // propagates invalidity to descendants (e.g., $Extend children).
-    let resolver = PathResolver::build(index, false);
-
-    // Closure wraps the free helper `resolve_parent_compact_idx` so the
-    // typed `ParentFrs`/`Frs` signature is enforced at every call site
-    // (own↔parent swap becomes a compile error).  Keeping the helper
-    // free-standing also keeps `build_compact_index` under the
-    // clippy::too_many_lines budget.
-    let resolve_parent = |parent_frs: uffs_mft::ParentFrs, own_frs: uffs_mft::Frs| -> u32 {
-        resolve_parent_compact_idx(index, parent_frs, own_frs)
-    };
-
-    // Phase 1: build primary compact records (parallel).
-    let mut records: Vec<CompactRecord> = index
-        .records
-        .par_iter()
-        .enumerate()
-        .map(|(idx, record)| {
-            // Skip invalid records (system metafiles + descendants).
-            if !resolver.is_valid_idx(idx) {
-                return CompactRecord::default();
-            }
-
-            let name_ref = &record.first_name.name;
-            let parent_idx = resolve_parent(record.first_name.parent_frs, record.frs);
-
-            CompactRecord {
-                size: record.first_stream.size.length,
-                allocated: record.first_stream.size.allocated,
-                treesize: record.treesize,
-                tree_allocated: record.tree_allocated,
-                created: record.stdinfo.created,
-                modified: record.stdinfo.modified,
-                accessed: record.stdinfo.accessed,
-                name_offset: name_ref.offset,
-                flags: record.stdinfo.flags,
-                parent_idx,
-                descendants: record.descendants,
-                name_len: name_ref.length(),
-                extension_id: name_ref.extension_id(),
-                path_len: 0,
-                name_first_byte: index
-                    .names
-                    .get(name_ref.offset as usize)
-                    .copied()
-                    .unwrap_or(0),
-                _pad: [0; 1],
-            }
-        })
-        .collect();
-
-    // Phase 2+3: expand hardlinks and ADS (sequential — rare, <1% of records).
-    let mut names = index.names.clone();
-    let expanded = expand_links_and_ads(index, &resolver, &resolve_parent, &mut names);
-    records.extend(expanded);
-
-    // Phase 4: compute path_len (in characters) for every record via
-    // top-down BFS.  path_len = char count of "C:\dir\name".
-    compute_path_lengths(&mut records, &names, drive_letter);
-
-    let compact_elapsed = compact_start.elapsed().as_millis();
-
-    // Try live $UpCase from the NTFS volume; fall back to compiled-in default.
-    let fold = resolve_case_fold(drive_letter);
-
-    let tri_start = Instant::now();
-    let trigram = TrigramIndex::build(&records, &names, fold);
-    let tri_elapsed = tri_start.elapsed().as_millis();
-
-    // Build children CSR index from parent_idx (two-pass: count + scatter).
-    let children = ChildrenIndex::build(&records);
-
-    // Copy extension name table from MftIndex (Arc<str> → Box<str>).
-    let mut ext_names: Vec<Box<str>> = index
-        .extensions
-        .names
-        .iter()
-        .map(|arc| Box::from(arc.as_ref()))
-        .collect();
-
-    let ext_t0 = Instant::now();
-    let ext_index = ExtensionIndex::build(&records);
-    let ext_build_ms = ext_t0.elapsed().as_millis();
-    tracing::info!(
-        drive = %drive_letter,
-        entries = ext_index.total_entries(),
-        build_ms = ext_build_ms,
-        "ExtensionIndex built"
-    );
-
-    shrink_compact_vecs(drive_letter, &mut records, &mut names, &mut ext_names);
-
-    // Phase 8: clone the FRS → mft_idx mapping off the transient
-    // `MftIndex` before it goes out of scope.  In the primary
-    // `build_compact_index` path compact_idx == mft_idx (records
-    // are produced 1:1 by `index.records.par_iter().enumerate()`),
-    // so `frs_to_idx` is exactly the FRS → compact_idx mapping the
-    // surgical-patch path needs.  Hardlink / ADS-expanded records
-    // append at the END with the same FRS but higher compact_idx;
-    // those secondary slots are not addressable from journal events
-    // (USN events reference primary FRS) so the primary mapping is
-    // sufficient.  `uffs_mft::NO_ENTRY == u32::MAX` matches the
-    // sentinel `frs_to_compact` uses for unmapped slots.
-    let mut compact_index = DriveCompactIndex {
-        letter: drive_letter,
-        records: ColumnStorage::from_vec(records),
-        names: ColumnStorage::from_vec(names),
-        trigram,
-        children,
-        ext_index,
-        fold,
-        ext_names,
-        source: IndexSource::MftFile(std::path::PathBuf::from(format!("{drive_letter}:"))),
-        source_epoch: index.build_epoch,
-        bloom: None,
-        path_trie: None,
-        frs_to_compact: index.frs_to_idx.clone(),
-        // Freshly built from the MFT — base CSR indexes are authoritative,
-        // no overlay yet. apply_usn_patch (Phase 2b) starts the delta.
-        delta: None,
-    };
-
-    // Phase 4: populate bloom + path_trie from the freshly-built
-    // index.  These are needed for the search-skip pre-check
-    // (Commit F) and serialised into the v9+ cache (Commit D).
-    let bloom = compact_index.build_bloom();
-    let path_trie = compact_index.build_path_trie();
-    compact_index.bloom = Some(bloom);
-    compact_index.path_trie = Some(path_trie);
-
-    (compact_index, compact_elapsed, tri_elapsed)
-}
-
-/// Shrink all growable Vecs to exact fit after compact index build.
-///
-/// Reclaims capacity slack from the doubling growth strategy used during
-/// construction.  Saves ~500 MB across 7 drives.
-fn shrink_compact_vecs(
-    drive_letter: uffs_mft::platform::DriveLetter,
-    records: &mut Vec<CompactRecord>,
-    names: &mut Vec<u8>,
-    ext_names: &mut Vec<Box<str>>,
-) {
-    let pre = records.capacity() * size_of::<CompactRecord>() + names.capacity();
-    records.shrink_to_fit();
-    names.shrink_to_fit();
-    ext_names.shrink_to_fit();
-    let post = records.capacity() * size_of::<CompactRecord>() + names.capacity();
-    let reclaimed_mb = pre.saturating_sub(post) / (1024 * 1024);
-    if reclaimed_mb > 0 {
-        tracing::info!(
-            drive = %drive_letter,
-            reclaimed_mb,
-            "shrink_to_fit reclaimed memory"
-        );
-    }
-}
-
-/// Cache TTL in seconds (4 hours — same as Windows CLI).
-///
-/// USN Journal handles incremental freshness; this is a safety-net full rescan.
-pub(crate) const INDEX_TTL_SECONDS: u64 = 14400;
-
-// ── Live $UpCase resolution ──────────────────────────────────────────
-
-/// Try to read the live `$UpCase` table from the NTFS volume for
-/// `drive_letter`. On success, log the result at `INFO` and any diffs
-/// from the compiled-in default at `WARN`. On failure, log at `WARN`
-/// and fall back to [`CaseFold::default_table()`].
-pub(crate) fn resolve_case_fold(
-    drive_letter: uffs_mft::platform::DriveLetter,
-) -> uffs_text::case_fold::CaseFold {
-    let live_table = match uffs_mft::platform::upcase::read_upcase_table(drive_letter) {
-        Ok(table) => table,
-        Err(err) => {
-            tracing::warn!(
-                drive = %drive_letter,
-                error = %err,
-                "$UpCase live read failed — falling back to compiled-in default table"
-            );
-            return uffs_text::case_fold::CaseFold::default_table();
-        }
-    };
-
-    // Leak the box to get a `&'static [u16]` for CaseFold::from_ntfs.
-    let live_fold = uffs_text::case_fold::CaseFold::from_ntfs(Box::leak(live_table));
-    log_upcase_comparison(drive_letter, &live_fold);
-    live_fold
-}
-
-/// Log the comparison between live and compiled-in `$UpCase` tables.
-fn log_upcase_comparison(
-    drive_letter: uffs_mft::platform::DriveLetter,
-    live_fold: &uffs_text::case_fold::CaseFold,
-) {
-    let default = uffs_text::case_fold::CaseFold::default_table();
-    let diffs = default.diff(live_fold);
-
-    if diffs.is_empty() {
-        tracing::info!(
-            drive = %drive_letter,
-            "$UpCase loaded from live volume — identical to compiled-in default"
-        );
-        return;
-    }
-
-    tracing::info!(
-        drive = %drive_letter,
-        diff_count = diffs.len(),
-        "$UpCase loaded from live volume — differs from compiled-in default"
-    );
-    for diff in &diffs {
-        tracing::warn!(
-            drive = %drive_letter,
-            codepoint = format_args!("U+{:04X}", diff.codepoint),
-            default = format_args!("U+{:04X}", diff.default_maps_to),
-            live = format_args!("U+{:04X}", diff.live_maps_to),
-            "$UpCase diff"
-        );
     }
 }
